// desktop/src/ipc/generated/WhoamiReport.ts
// SPDX-License-Identifier: FSL-1.1-Apache-2.0
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CliDefaultTarget } from "./CliDefaultTarget";
import type { Identity } from "./Identity";

/**
 * What whoami reports.
 */
export type WhoamiReport = { identity: Identity, cliDefault: CliDefaultTarget, };